ReactOS 0.4.1 was already released! Tag the branch.

+++ reactos/drivers/filesystems/ext2/inc/linux/atomic.h
@@ -8,7 +8,7 @@
 //
 
 typedef struct {
-    volatile LONG  counter;
+    volatile int counter;
 } atomic_t;
 
 #define ATOMIC_INIT(i)	(i)
